refactor(core): separate legend placement from per-guide builders
Split packages/core legend construction so multi-legend placement
orchestration and the public re-export surface stay in legend.ts, while
types, text helpers, and discrete/ramp/steps builders live in
legend-build.ts. Add characterization tests for domain order, auto bottom
degradation, block stacking, and post-placement overflow checks.
